+++ middleware/multerUpload.js
@@ -1,0 +1,18 @@
+const multer = require('multer')
+
+const upload = multer({
+    limits: {
+        fileSize: 1000000
+    },
+    fileFilter(req, file, cb) {
+        console.log("file multer")
+        if (!file.originalname.match(/\.(jpeg|png|jpg)$/))
+            return cb(new Error('Please Upload an Image'))
+        cb(undefined, true)
+    }
+})
+
+
+module.exports = {
+    upload
+}
